+++ e2e/cases/source-map/src/index.css
@@ -1,0 +1,3 @@
+body {
+  color: red;
+}
