+++ e2e/cases/source-map/src/index.js
@@ -2,6 +2,7 @@
 import React from 'react';
 import { render } from 'react-dom';
 import App from './App';
+import './index.css';
 
 window.aa = 2;
